fix: LayerButton
