+++ app/routes/_index/features/LayerButton/index.ts
@@ -1,0 +1,1 @@
+export { LayerButton as layer } from "./LayerButton";
