--- app/routes/_index/features/layer/index.ts
@@ -1,1 +1,0 @@
-export { layer } from "./layer";
